refactor endpoint network response handling
